requirements.txt
clarifai
streamlit
Pillow
requests
python-dotenv